staging: fsl-dpaa2/eth: Fix adding FS rule at specific location
When inserting a flow steering rule through ethtool, the user
may specify a location index. Ethtool manual says the expected
behaviour is to overwrite any rule present in that location.

What currently happens is that the driver passes on the location
to MC with no further verifications and if there's already
a rule present at that index, MC returns an error.

Fix this by making sure the location is available, and if not
free the current entry before inserting the new one.

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@nxp.com>